generate posts after edit

+++ server/delete.js
@@ -17,6 +17,16 @@ module.exports = async (req, res) => {
 
   const db = await sqlite.open(path.resolve(__dirname, "..", "posts.db"));
 
+  const dbPost = await db.get(
+    `SELECT id, draft FROM posts WHERE id = ?1 AND draft = 1 LIMIT 1`,
+    { 1: existingPostId }
+  );
+
+  if (!dbPost) {
+    res.statusCode = 400;
+    return `make post a draft before removing it`;
+  }
+
   await db.run(`DELETE FROM posts WHERE id = ?1`, {
     1: existingPostId
   });
